+++ background_scripts/actions.js
@@ -665,13 +665,16 @@ Actions = (function() {
 
   _.editWithVim = function() {
     var xhr = new XMLHttpRequest();
-    xhr.open('POST', 'http://127.0.0.1:8001');
-    xhr.addEventListener('readystatechange', function() {
+    xhr.open('POST', 'http://127.0.0.1:' + Settings.vimport);
+    xhr.onreadystatechange = function() {
       if (xhr.readyState === 4 && xhr.status === 200) {
         callback({type: 'editWithVim', text: xhr.responseText});
       }
-    });
-    xhr.send('' + request.text);
+    };
+    xhr.send(JSON.stringify({
+      command: Settings.vimcommand,
+      data: '' + (request.text || '')
+    }));
   };
 
   _.httpRequest = function() {